requirements.txt
requests
torch
transformers
biopython
sentencepiece
hf_xet